test(kvstoreentry): write tests for KV Store entries

+++ pkg/commands/kvstoreentry/delete.go
@@ -6,6 +6,7 @@ import (
 	"github.com/fastly/go-fastly/v8/fastly"
 
 	"github.com/fastly/cli/pkg/cmd"
+	fsterr "github.com/fastly/cli/pkg/errors"
 	"github.com/fastly/cli/pkg/global"
 	"github.com/fastly/cli/pkg/manifest"
 	"github.com/fastly/cli/pkg/text"
@@ -14,6 +15,8 @@ import (
 // DeleteCommand calls the Fastly API to delete an kv store.
 type DeleteCommand struct {
 	cmd.Base
+	cmd.JSONOutput
+
 	manifest manifest.Data
 	Input    fastly.DeleteKVStoreKeyInput
 }
@@ -27,19 +30,43 @@ func NewDeleteCommand(parent cmd.Registerer, g *global.Data, m manifest.Data) *D
 		manifest: m,
 	}
 	c.CmdClause = parent.Command("delete", "Delete a key")
+
+	// Required.
+	c.CmdClause.Flag("key", "Key name").Short('k').Required().StringVar(&c.Input.Key)
 	c.CmdClause.Flag("store-id", "Store ID").Short('s').Required().StringVar(&c.Input.ID)
-	c.CmdClause.Flag("key-name", "Key name").Short('k').Required().StringVar(&c.Input.Key)
+
+	// Optional.
+	c.RegisterFlagBool(c.JSONFlag()) // --json
+
 	return &c
 }
 
 // Exec invokes the application logic for the command.
 func (c *DeleteCommand) Exec(_ io.Reader, out io.Writer) error {
+	if c.Globals.Verbose() && c.JSONOutput.Enabled {
+		return fsterr.ErrInvalidVerboseJSONCombo
+	}
+
 	err := c.Globals.APIClient.DeleteKVStoreKey(&c.Input)
 	if err != nil {
 		c.Globals.ErrLog.Add(err)
 		return err
 	}
 
-	text.Success(out, "Deleted key %s from KV Store ID %s", c.Input.Key, c.Input.ID)
+	if c.JSONOutput.Enabled {
+		o := struct {
+			Key     string `json:"key"`
+			ID      string `json:"id"`
+			Deleted bool   `json:"deleted"`
+		}{
+			c.Input.Key,
+			c.Input.ID,
+			true,
+		}
+		_, err := c.WriteJSON(out, o)
+		return err
+	}
+
+	text.Success(out, "Deleted key %s from KV Store %s", c.Input.Key, c.Input.ID)
 	return nil
 }
